requirements.txt
django
mysqlclient
whitenoise
djangorestframework
django-tables2
PrettyTable
bootstrap4